requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
joblib==1.4.2
lxml==5.3.0
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
